🦄 refactor: modularize AI and terminal managers into base classes and dedicated helper modules
