+++ package.json
@@ -94,10 +94,6 @@
         "key": "alt+v",
         "mac": "alt+v",
         "command": "lynx-keymap.openGitlabPanel"
-      },{
-        "key": "ctrl+4",
-        "mac": "cmd+4",
-        "command": "lynx-keymap.openMySkillsPanel"
       },
       {
         "key"    : "alt+b",
@@ -434,9 +430,13 @@
       },
       
       {
+        "key"    : "ctrl+4",
+        "mac"    : "cmd+4",
+        "command": "workbench.view.extensions"
+      },{
         "key"    : "ctrl+3",
         "mac"    : "cmd+3",
-        "command": "workbench.view.extensions"
+        "command": "lynx-keymap.openMySkillsPanel"
       },{
         "key"    : "ctrl+2",
         "mac"    : "cmd+2",
