+++ src/extension.ts
@@ -1,115 +1,74 @@
 import * as vscode from 'vscode';
-import { AICommandsManager, AIToggleManager, BottomTerminalManager, DebugManager, TerminalManager, WordWrapManager } from './keymaps';
-import { STORAGE_KEYS, PANEL_POSITIONS, LOG_PREFIX } from './shared/constants';
-import { promptInstallAtmExtension, promptInstallMySkillsExtension } from './notifications/with-buttons';
+import { EditorDetector }              from './keymaps/ai/detector';
+import { AICommandsManager }           from './keymaps/ai/commands-manager';
+import { AIToggleManager }             from './keymaps/ai/toggle-manager';
+import { TerminalManager }             from './keymaps/terminal/side-panel';
+import { BottomTerminalManager }       from './keymaps/terminal/bottom-panel';
+import { DebugManager }                from './debug/panel';
+import { WordWrapManager }             from './wordwrap/manager';
+import { recoverSidePanelState }       from './keymaps/terminal/startup-recovery';
+import { promptInstallExtension }      from './notifications/install-prompt';
 
-let aiManager:             AICommandsManager     | undefined;
-let terminalManager:       TerminalManager       | undefined;
-let bottomTerminalManager: BottomTerminalManager | undefined;
-let aiToggleManager:       AIToggleManager       | undefined;
-let wordWrapManager:       WordWrapManager       | undefined;
-let debugManager:          DebugManager          | undefined;
-let startupTimeoutId:      NodeJS.Timeout        | undefined;
+const managers: Array<{ name: string; ref: vscode.Disposable | undefined }> = [];
 
 export async function activate(context: vscode.ExtensionContext) {
-  aiManager             = new AICommandsManager();
-  terminalManager       = new TerminalManager();
-  bottomTerminalManager = new BottomTerminalManager();
-  aiToggleManager       = new AIToggleManager(aiManager);
-  wordWrapManager       = new WordWrapManager();
-  debugManager          = new DebugManager();
+  const detector           = new EditorDetector();
+  const aiManager          = new AICommandsManager(detector);
+  const aiToggleManager    = new AIToggleManager(detector);
+  const terminalManager    = new TerminalManager();
+  const bottomTerminalMgr  = new BottomTerminalManager();
+  const wordWrapManager    = new WordWrapManager();
+  const debugManager       = new DebugManager();
+
+  managers.push(
+    { name: 'aiManager',         ref: aiManager         },
+    { name: 'aiToggleManager',   ref: aiToggleManager   },
+    { name: 'terminalManager',   ref: terminalManager   },
+    { name: 'bottomTerminalMgr', ref: bottomTerminalMgr },
+    { name: 'wordWrapManager',   ref: wordWrapManager   },
+    { name: 'debugManager',      ref: debugManager      },
+  );
 
   aiManager.registerCommands(context);
-  terminalManager.registerCommands(context);
-  bottomTerminalManager.registerCommands(context);
   aiToggleManager.registerCommands(context);
+  terminalManager.registerCommands(context);
+  bottomTerminalMgr.registerCommands(context);
   wordWrapManager.registerCommands(context);
   debugManager.registerCommands(context);
 
-  // Pre-detect the editor so the first keypress is instant.
   void aiManager.warmup().catch(error => {
-    console.warn(`${LOG_PREFIX} AI detection warmup failed:`, error);
+    console.warn(`[lynx-keymap] AI detection warmup failed:`, error);
   });
 
-  const gitlabPanelCommand = vscode.commands.registerCommand('lynx-keymap.openGitlabPanel', async () => {
-    const atmExtension = vscode.extensions.getExtension('bastndev.atm');
-    if (atmExtension) {
-      if (!atmExtension.isActive) {
-        await atmExtension.activate();
-      }
-      void vscode.commands.executeCommand('workbench.view.extension.gitlab-panel');
-    } else {
-      void promptInstallAtmExtension();
-    }
-  });
-  context.subscriptions.push(gitlabPanelCommand);
-
-  const mySkillsPanelCommand = vscode.commands.registerCommand('lynx-keymap.openMySkillsPanel', async () => {
-    const mySkillsExtension = vscode.extensions.getExtension('bastndev.my-skills');
-    if (mySkillsExtension) {
-      if (!mySkillsExtension.isActive) {
-        await mySkillsExtension.activate();
-      }
-      void vscode.commands.executeCommand('myskills-panel.focus');
-    } else {
-      void promptInstallMySkillsExtension();
-    }
-  });
-  context.subscriptions.push(mySkillsPanelCommand);
-
-  // Read previous position BEFORE resetting — needed for startup cleanup below.
-  const prevPosition = context.workspaceState.get<string>(STORAGE_KEYS.PANEL_POSITION);
-
-  await context.workspaceState.update(STORAGE_KEYS.PANEL_POSITION,           undefined);
-  await context.globalState.update(STORAGE_KEYS.ORIGINAL_TABS_ENABLED,       undefined);
-  await context.globalState.update(STORAGE_KEYS.ORIGINAL_PANEL_SHOW_LABELS,  undefined);
-  await context.globalState.update(STORAGE_KEYS.ORIGINAL_TABS_LOCATION,      undefined);
-
-  // If the terminal was left in side-panel mode, close the auxiliary bar on startup.
-  // Some editors restore it aggressively, so we retry at staggered intervals.
-  // closeAuxiliaryBar is idempotent — safe to call even when already closed.
-  if (prevPosition === PANEL_POSITIONS.LEFT) {
-    const closeAuxBar = async () => {
-      try {
-        await vscode.commands.executeCommand('workbench.action.closeAuxiliaryBar');
-      } catch (error) {
-        console.debug(`${LOG_PREFIX} Auxiliary bar cleanup skipped:`, error);
-      }
-    };
-
-    setTimeout(closeAuxBar, 300);
-    setTimeout(closeAuxBar, 800);
-    setTimeout(closeAuxBar, 1600);
-    startupTimeoutId = setTimeout(async () => {
-      try {
-        await closeAuxBar();
-      } finally {
-        startupTimeoutId = undefined;
-      }
-    }, 3000);
-  }
-}
-
-export async function deactivate() {
-  if (startupTimeoutId) {
-    clearTimeout(startupTimeoutId);
-    startupTimeoutId = undefined;
-  }
-
-  const managers: Array<{ name: string; ref: vscode.Disposable | undefined }> = [
-    { name: 'aiManager',             ref: aiManager             },
-    { name: 'terminalManager',       ref: terminalManager       },
-    { name: 'bottomTerminalManager', ref: bottomTerminalManager },
-    { name: 'aiToggleManager',       ref: aiToggleManager       },
-    { name: 'wordWrapManager',       ref: wordWrapManager       },
-    { name: 'debugManager',          ref: debugManager          },
+  // ─── Panel commands (GitLab + MySkills) ────────────────────────────────────
+  const panelCommands: Array<{ cmd: string; extId: string; focusCmd: string }> = [
+    { cmd: 'lynx-keymap.openGitlabPanel',   extId: 'bastndev.atm',       focusCmd: 'workbench.view.extension.gitlab-panel' },
+    { cmd: 'lynx-keymap.openMySkillsPanel',  extId: 'bastndev.my-skills', focusCmd: 'myskills-panel.focus' },
   ];
 
+  for (const { cmd, extId, focusCmd } of panelCommands) {
+    const disposable = vscode.commands.registerCommand(cmd, async () => {
+      const ext = vscode.extensions.getExtension(extId);
+      if (ext) {
+        if (!ext.isActive) { await ext.activate(); }
+        void vscode.commands.executeCommand(focusCmd);
+      } else {
+        void promptInstallExtension(extId);
+      }
+    });
+    context.subscriptions.push(disposable);
+  }
+
+  // ─── Startup recovery ──────────────────────────────────────────────────────
+  await recoverSidePanelState(context);
+}
+
+export function deactivate(): void {
   for (const { name, ref } of managers) {
     try {
       ref?.dispose();
     } catch (error) {
-      console.error(`${LOG_PREFIX} Error disposing ${name}:`, error);
+      console.error(`[lynx-keymap] Error disposing ${name}:`, error);
     }
   }
 }
