+++ src/keymaps/terminal/constants.ts
@@ -1,0 +1,7 @@
+// ─── Panel Config ─────────────────────────────────────────────────────────────
+export const TERMINAL_CONFIG  = 'terminal.integrated';
+export const WORKBENCH_CONFIG = 'workbench';
+
+// VS Code may briefly re-assert the auxiliary bar after panel commands;
+// this delay lets the layout settle before a follow-up close takes effect.
+export const LAYOUT_SETTLE_MS = 150;
