--- src/notifications/with-buttons.ts
@@ -1,26 +1,0 @@
-import * as vscode from 'vscode';
-import { getTranslation } from './info';
-
-export async function promptInstallAtmExtension(): Promise<void> {
-  const installAction = await getTranslation('ATM.notification.install.action');
-  const selection = await vscode.window.showInformationMessage(
-    await getTranslation('ATM.notification.install.required'),
-    installAction
-  );
-
-  if (selection === installAction) {
-    void vscode.commands.executeCommand('workbench.extensions.installExtension', 'bastndev.atm');
-  }
-}
-
-export async function promptInstallMySkillsExtension(): Promise<void> {
-  const installAction = await getTranslation('ATM.notification.install.action');
-  const selection = await vscode.window.showInformationMessage(
-    await getTranslation('ATM.notification.install.required'),
-    installAction
-  );
-
-  if (selection === installAction) {
-    void vscode.commands.executeCommand('workbench.extensions.installExtension', 'bastndev.my-skills');
-  }
-}
